requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
langchain-chroma
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
altair==4.2.0
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
cohere
pdf2image
google-search-results
transformers
unstructured
langchain_cohere
guardrails-ai
langchain_experimental
langchain-google-genai
validators
youtube_transcript_api
pytube
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube